src/main/java/racing/domain/RacingCars.java
package racing.domain;

import java.util.List;
import java.util.Objects;

public class RacingCars {
    private final List<Car> cars;

    public RacingCars(List<Car> cars) {
        this.cars = cars;
    }

    public void race(CarMovable movable) {
        this.cars.forEach(car -> car.move(movable));
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;

        RacingCars cars1 = (RacingCars) o;

        return Objects.equals(cars, cars1.cars);
    }

    @Override
    public int hashCode() {
        return cars != null ? cars.hashCode() : 0;
    }
}
